requirements.txt
numpy>=1.19.0
opencv-python>=4.5.1
pillow>=8.0.0
gradio>=3.0.0
torch>=1.7.1
torchvision>=0.8.2